Task 44: Отрефакторены тесты с применением pytest_plugins
