+++ theater-async-api/conftest.py
@@ -1,0 +1,5 @@
+pytest_plugins = [
+    "tests.functional.fixtures.test_fixtures_elasticsearch",
+    "tests.functional.fixtures.test_fixtures_redis",
+    "tests.functional.fixtures.test_fixtures_http",
+]
